feat(hal/cortex_m): promote all qemu/cortex components to cortex_m
Move SystemTick, SystemTickTimerService, Reset, EventDispatcherCortex,
and Semihosting from hal/qemu/cortex/ to hal/cortex_m/ so any Cortex-M
HAL (not just QEMU) can use them without reimplementing locally.

Key changes:
- Reset now implements hal::Reset (takes const char* resetReason)
- EventDispatcherCortexWorker switches base to EventDispatcherWorkerImpl
  and uses DSB+WFE/SEV for proper ARM sleep/wakeup instead of bare WFI
- hal.qemu.cortex becomes an INTERFACE library (no sources remain)
- All hal/qemu/default_init and test consumers updated to new paths

Closes #100, #101, #102
