+++ hal/cortex_m/Semihosting.cpp
@@ -1,4 +1,4 @@
-#include "hal/qemu/cortex/Semihosting.hpp"
+#include "hal/cortex_m/Semihosting.hpp"
 
 namespace hal::cortex
 {
