--- hal/qemu/cortex/EventDispatcherCortex.cpp
@@ -1,12 +1,0 @@
-#include "hal/qemu/cortex/EventDispatcherCortex.hpp"
-
-namespace hal::cortex
-{
-    void EventDispatcherCortexWorker::RequestExecution()
-    {}
-
-    void EventDispatcherCortexWorker::Idle()
-    {
-        __asm volatile("wfi");
-    }
-}
