--- hal/qemu/cortex/Reset.hpp
@@ -1,13 +1,0 @@
-#ifndef HAL_QEMU_CORTEX_RESET_HPP
-#define HAL_QEMU_CORTEX_RESET_HPP
-
-namespace hal::cortex
-{
-    class Reset
-    {
-    public:
-        void ResetModule();
-    };
-}
-
-#endif
