+++ hal/qemu/default_init/TracerOnSemihostingInfrastructure.cpp
@@ -1,5 +1,5 @@
 #include "hal/qemu/default_init/TracerOnSemihostingInfrastructure.hpp"
-#include "hal/qemu/cortex/Semihosting.hpp"
+#include "hal/cortex_m/Semihosting.hpp"
 #include "services/tracer/GlobalTracer.hpp"
 
 namespace bringup
